Pin two inferred types on the game-detail panes
The tab-fallback ternary and the field bar's offset both left the
type-checker to bridge CGFloat/Double or infer an enum from the other
branch. Naming them costs nothing and keeps the build cheap — the
SwipeSafeButtonStyle note's rule.

+++ sports/Features/GameDetail/LiveSituationCard.swift
@@ -71,7 +71,7 @@ struct LiveSituationCard: View {
     private var field: some View {
         VStack(spacing: Spacing.xs) {
             GeometryReader { geo in
-                let fraction = situation.fieldPosition ?? 0
+                let fraction = CGFloat(situation.fieldPosition ?? 0)
                 ZStack(alignment: .leading) {
                     Capsule()
                         .fill(Color.divider)
